refactoring of skill access methods
