+++ src/ai/susi/mind/SusiSkill.java
@@ -25,19 +25,8 @@ import java.io.FileNotFoundException;
 import java.io.FileReader;
 import java.io.IOException;
 import java.nio.file.FileSystems;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.nio.file.attribute.BasicFileAttributes;
-import java.nio.file.attribute.FileTime;
-import java.util.Collections;
-import java.util.Comparator;
 import java.util.LinkedHashSet;
-import java.util.List;
-import java.util.Map;
 import java.util.Set;
-import java.util.Date;
-import java.text.DateFormat;
 
 import org.jfree.util.Log;
 import org.json.JSONArray;
@@ -46,10 +35,8 @@ import org.json.JSONObject;
 import org.json.JSONTokener;
 
 import ai.susi.DAO;
-import ai.susi.json.JsonTray;
 import ai.susi.mind.SusiAction.SusiActionException;
 import ai.susi.mind.SusiInference.Type;
-import ai.susi.tools.DateParser;
 
 /**
  * A susi skill is a set of intents.
@@ -502,612 +489,6 @@ public class SusiSkill {
             parent.put("cues", cues);
         }
         cues.put(utterance);
-    }
-    
-    /**
-     * For some strange reason the skill name is requested here in lowercase, while the name may also be uppercase
-     * this should be fixed in the front-end, however we implement a patch here to circumvent the problem if possible
-     * Another strange effect is, that some file systems do match lowercase with uppercase (like in windows),
-     * so testing skill.exists() would return true even if the name does not exist exactly as given in the file system.
-     * @param languagepath
-     * @param skill_name
-     * @return the actual skill file if one exist or a skill file that is constructed from language and skill_name
-     */
-    public static File getSkillFileInLanguage(File languagepath, String skill_name, boolean null_if_not_found) {
-
-        String fn = skill_name + ".txt";
-        String[] list = languagepath.list();
-
-        // first try: the skill name may be same or similar to the skill file name
-        if(list !=null && list.length!=0){
-            for (String n: list) {
-                if (n.equals(fn) || n.toLowerCase().equals(fn)) {
-                    return new File(languagepath, n);
-                }
-            }
-        }
-
-        // second try: the skill name may be same or similar to the skill name within the skill description
-        // this is costly: we must parse the whole skill file
-        if(list !=null && list.length!=0){
-            for (String n: list) {
-                if (!n.endsWith(".txt") && !n.endsWith(".ezd")) continue;
-                File f = new File(languagepath, n);
-                try {
-                    SusiSkill.ID skillid = new SusiSkill.ID(f);
-                    SusiLanguage language = skillid.language();
-                    JSONObject json = SusiSkill.readLoTSkill(new BufferedReader(new FileReader(f)), language, skillid.skillpath, false);
-                    String sn = json.optString("skill_name");
-                    if (sn.equals(skill_name) || sn.toLowerCase().equals(skill_name) || sn.toLowerCase().replace(' ', '_').equals(skill_name)) {
-                        return new File(languagepath, n);
-                    }
-                } catch (JSONException | FileNotFoundException e) {
-                    continue;
-                }
-            }
-        }
-
-        // the final attempt is bad and may not succeed, but it's the only last thing left we could do.
-        return null_if_not_found ? null : new File(languagepath, fn);
-    }
-
-    /**
-     * the following method scans a given model for all files to see if it matches the skill name
-     * @param model a path to a model directory
-     * @param skill_name
-     * @return
-     */
-    public static File getSkillFileInModel(File model, String skill_name) {
-        String[] groups = model.list();
-        for (String group: groups) {
-            if (group.startsWith(".")) continue;
-            File gf = new File(model, group);
-            if (!gf.isDirectory()) continue;
-            String[] languages = gf.list();
-            for (String language: languages) {
-                if (language.startsWith(".")) continue;
-                File l = new File(gf, language);
-                if (!l.isDirectory()) continue;
-                File skill = getSkillFileInLanguage(l, skill_name, true);
-                if (skill != null) return skill;
-            }
-        }
-        return null;
-    }
-
-    public static JSONObject getSkillMetadata(String model, String group, String language, String skillname, int duration) {
-
-        JSONObject skillMetadata = new JSONObject(true)
-                .put("model", model)
-                .put("group", group)
-                .put("language", language);
-        File modelpath = new File(DAO.model_watch_dir, model);
-        File grouppath = new File(modelpath, group);
-        File languagepath = new File(grouppath, language);
-        File skillpath = getSkillFileInLanguage(languagepath, skillname, false);
-        DateFormat dateFormatType = DateParser.iso8601Format;
-        skillname = skillpath.getName().replaceAll(".txt", ""); // fixes the bad name (lowercased) to the actual right name
-
-        // default values
-        skillMetadata.put("developer_privacy_policy", JSONObject.NULL);
-        skillMetadata.put("descriptions",JSONObject.NULL);
-        skillMetadata.put("image", JSONObject.NULL);
-        skillMetadata.put("author", JSONObject.NULL);
-        skillMetadata.put("author_url", JSONObject.NULL);
-        skillMetadata.put("author_email", JSONObject.NULL);
-        skillMetadata.put("skill_name", JSONObject.NULL);
-        skillMetadata.put("protected", false);
-        skillMetadata.put("reviewed", false);
-        skillMetadata.put("editable", true);
-        skillMetadata.put("staffPick", false);
-        skillMetadata.put("systemSkill", false);
-        skillMetadata.put("terms_of_use", JSONObject.NULL);
-        skillMetadata.put("dynamic_content", false);
-        skillMetadata.put("examples", JSONObject.NULL);
-        skillMetadata.put("skill_rating", JSONObject.NULL);
-        skillMetadata.put("usage_count", 0);
-        skillMetadata.put("skill_tag", JSONObject.NULL);
-        skillMetadata.put("lastModifiedTime", dateFormatType.format(new Date(0)));
-        skillMetadata.put("creationTime", dateFormatType.format(new Date(0)));
-
-        // metadata
-        for (Map.Entry<SusiSkill.ID, SusiSkill> entry : DAO.susi.getSkillMetadata().entrySet()) {
-            SusiSkill skill = entry.getValue();
-            SusiSkill.ID skillid = entry.getKey();
-            if (skillid.hasModel(model) &&
-                    skillid.hasGroup(group) &&
-                    skillid.hasLanguage(language) &&
-                    skillid.hasName(skillname)) {
-
-                skillMetadata.put("skill_name", skill.getSkillName() ==null ? JSONObject.NULL: skill.getSkillName());
-                skillMetadata.put("protected", skill.getProtectedSkill());
-                skillMetadata.put("developer_privacy_policy", skill.getDeveloperPrivacyPolicy() ==null ? JSONObject.NULL:skill.getDeveloperPrivacyPolicy());
-                skillMetadata.put("descriptions", skill.getDescription() ==null ? JSONObject.NULL:skill.getDescription());
-                skillMetadata.put("image", skill.getImage() ==null ? JSONObject.NULL: skill.getImage());
-                skillMetadata.put("author", skill.getAuthor()  ==null ? JSONObject.NULL:skill.getAuthor());
-                skillMetadata.put("author_url", skill.getAuthorURL() ==null ? JSONObject.NULL:skill.getAuthorURL());
-                skillMetadata.put("author_email", skill.getAuthorEmail() ==null ? JSONObject.NULL:skill.getAuthorEmail());
-                skillMetadata.put("terms_of_use", skill.getTermsOfUse() ==null ? JSONObject.NULL:skill.getTermsOfUse());
-                skillMetadata.put("dynamic_content", skill.getDynamicContent());
-                skillMetadata.put("examples", skill.getExamples() ==null ? JSONObject.NULL: skill.getExamples());
-                skillMetadata.put("skill_rating", getSkillRating(model, group, language, skillname));
-                skillMetadata.put("supported_languages", getSupportedLanguages(model, group, language, skillname));
-                skillMetadata.put("reviewed", getSkillReviewStatus(model, group, language, skillname));
-                skillMetadata.put("editable", getSkillEditStatus(model, group, language, skillname));
-                skillMetadata.put("staffPick", isStaffPick(model, group, language, skillname));
-                skillMetadata.put("systemSkill", isSystemSkill(model, group, language, skillname));
-                skillMetadata.put("usage_count", getSkillUsage(model, group, language, skillname, duration));
-                skillMetadata.put("skill_tag", skillname);
-                skillMetadata.put("lastModifiedTime", getSkillModifiedTime(model, group, language, skillname));
-                skillMetadata.put("creationTime", getSkillCreationTime(model, group, language, skillname, skillpath));
-            }
-        }
-
-        // file attributes
-        BasicFileAttributes attr = null;
-        Path p = Paths.get(skillpath.getPath());
-        try {
-            attr = Files.readAttributes(p, BasicFileAttributes.class);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        if(attr!=null){
-            skillMetadata.put("lastAccessTime" , attr.lastAccessTime());
-        }
-        return skillMetadata;
-    }
-
-    // Function overloading - if duration parameter is not passed then use 7 as default value.
-    public static JSONObject getSkillMetadata(String model, String group, String language, String skillname) {
-        return getSkillMetadata(model, group, language, skillname, 7);
-    }
-
-    public static JSONObject getSkillRating(String model, String group, String language, String skillname) {
-        // rating
-        JsonTray skillRating = DAO.skillRating;
-        if (skillRating.has(model)) {
-            JSONObject modelName = skillRating.getJSONObject(model);
-            if (modelName.has(group)) {
-                JSONObject groupName = modelName.getJSONObject(group);
-                if (groupName.has(language)) {
-                    JSONObject languageName = groupName.getJSONObject(language);
-                    if (languageName.has(skillname)) {
-                        JSONObject skillName = languageName.getJSONObject(skillname);
-
-                        if (!skillName.has("stars")) {
-                            JSONObject newFiveStarRating = new JSONObject();
-                            newFiveStarRating.put("one_star", 0);
-                            newFiveStarRating.put("two_star", 0);
-                            newFiveStarRating.put("three_star", 0);
-                            newFiveStarRating.put("four_star", 0);
-                            newFiveStarRating.put("five_star", 0);
-                            newFiveStarRating.put("avg_star", 0);
-                            newFiveStarRating.put("total_star", 0);
-
-                            skillName.put("stars", newFiveStarRating);
-                        }
-                        return skillName;
-                    }
-                }
-            }
-        }
-        JSONObject newRating=new JSONObject();
-        newRating.put("negative", "0");
-        newRating.put("positive", "0");
-        newRating.put("feedback_count", 0);
-        newRating.put("bookmark_count", 0);
-
-        JSONObject newFiveStarRating=new JSONObject();
-        newFiveStarRating.put("one_star", 0);
-        newFiveStarRating.put("two_star", 0);
-        newFiveStarRating.put("three_star", 0);
-        newFiveStarRating.put("four_star", 0);
-        newFiveStarRating.put("five_star", 0);
-        newFiveStarRating.put("avg_star", 0);
-        newFiveStarRating.put("total_star", 0);
-
-        newRating.put("stars", newFiveStarRating);
-
-        return newRating;
-    }
-
-    public static JSONArray getSupportedLanguages(String model, String group, String language, String skillname) {
-        // rating
-        JsonTray skillRating = DAO.skillSupportedLanguages;
-        if (skillRating.has(model)) {
-            JSONObject modelName = skillRating.getJSONObject(model);
-            if (modelName.has(group)) {
-                JSONArray groupName = modelName.getJSONArray(group);
-
-                for (int i = 0; i < groupName.length(); i++) {
-                    JSONArray supportedLanguages = groupName.getJSONArray(i);
-                    for (int j = 0; j < supportedLanguages.length(); j++) {
-                        JSONObject languageObject = supportedLanguages.getJSONObject(j);
-                        String supportedLanguage = languageObject.get("language").toString();
-                        String skillName = languageObject.get("name").toString();
-                        if (supportedLanguage.equalsIgnoreCase(language) && skillName.equalsIgnoreCase(skillname)) {
-                            return supportedLanguages;
-                        }
-                    }
-                }
-            }
-        }
-        JSONArray supportedLanguages = new JSONArray();
-        JSONObject languageObject = new JSONObject();
-        languageObject.put("language", language);
-        languageObject.put("name", skillname);
-        supportedLanguages.put(languageObject);
-        return supportedLanguages;
-    }
-
-    public static String getSkillModifiedTime(String model_name, String group_name, String language_name, String skill_name) {
-        // SKill Info
-        JsonTray skillInfo = DAO.skillInfo;
-        JSONObject modelName = new JSONObject();
-        JSONObject groupName = new JSONObject();
-        JSONObject languageName = new JSONObject();
-        JSONObject skillName = new JSONObject();
-        DateFormat dateFormatType = DateParser.iso8601Format;
-        String skillModifiedTime = dateFormatType.format(new Date(0));
-
-        if (skillInfo.has(model_name)) {
-            modelName = skillInfo.getJSONObject(model_name);
-            if (modelName.has(group_name)) {
-                groupName = modelName.getJSONObject(group_name);
-                if (groupName.has(language_name)) {
-                    languageName = groupName.getJSONObject(language_name);
-                    if (languageName.has(skill_name)) {
-                        skillName = languageName.getJSONObject(skill_name);
-                        if (skillName.has("lastModifiedTime")) {
-                            skillModifiedTime = skillName.getString("lastModifiedTime");
-                        } else {
-                            skillName.put("lastModifiedTime", skillModifiedTime);
-                        }
-                        return skillModifiedTime;
-                    }
-                }
-            }
-        }
-
-        skillName.put("lastModifiedTime",skillModifiedTime);
-        languageName.put(skill_name, skillName);
-        groupName.put(language_name, languageName);
-        modelName.put(group_name, groupName);
-        skillInfo.put(model_name, modelName, true);
-        return skillModifiedTime;
-    }
-
-    public static String getSkillCreationTime(String model_name, String group_name, String language_name, String skill_name, File skill_path) {
-        // Skill Info
-        JsonTray skillInfo = DAO.skillInfo;
-        JSONObject modelName = new JSONObject();
-        JSONObject groupName = new JSONObject();
-        JSONObject languageName = new JSONObject();
-        JSONObject skillName = new JSONObject();
-        DateFormat dateFormatType = DateParser.iso8601Format;
-        String skillCreationTime = dateFormatType.format(new Date(0));
-
-        if (skillInfo.has(model_name)) {
-            modelName = skillInfo.getJSONObject(model_name);
-            if (modelName.has(group_name)) {
-                groupName = modelName.getJSONObject(group_name);
-                if (groupName.has(language_name)) {
-                    languageName = groupName.getJSONObject(language_name);
-                    if (languageName.has(skill_name)) {
-                        skillName = languageName.getJSONObject(skill_name);
-                        if (skillName.has("creationTime")) {
-                            skillCreationTime = skillName.getString("creationTime");
-                        } else {
-                            skillCreationTime = getFileCreationTime(skill_path);
-                            skillName.put("creationTime", skillCreationTime );
-                        }
-                        return skillCreationTime;
-                    }
-                }
-            }
-        }
-
-        skillCreationTime = getFileCreationTime(skill_path);
-        skillName.put("creationTime",skillCreationTime);
-        languageName.put(skill_name, skillName);
-        groupName.put(language_name, languageName);
-        modelName.put(group_name, groupName);
-        skillInfo.put(model_name, modelName, true);
-        return skillCreationTime;
-    }
-
-    public static String getFileCreationTime(File skill_path) {
-
-        DateFormat dateFormatType = DateParser.iso8601Format;
-        String skillCreationTime = dateFormatType.format(new Date(0));
-
-        BasicFileAttributes attr = null;
-        Path p = Paths.get(skill_path.getPath());
-        try {
-            attr = Files.readAttributes(p, BasicFileAttributes.class);
-        } catch (IOException e) {
-            e.printStackTrace();
-        }
-        if(attr!=null){
-            FileTime fileCreationTime = attr.creationTime();
-            skillCreationTime = dateFormatType.format(new Date(fileCreationTime.toMillis()));
-        }
-        return skillCreationTime;
-    }
-
-    public static void sortByAvgStar(List<JSONObject> jsonValues, boolean ascending) {
-        // Get skills based on ratings
-        Collections.sort(jsonValues, new Comparator<JSONObject>() {
-            @Override
-            public int compare(JSONObject a, JSONObject b) {
-                Object valA, valB;
-                int result=0;
-
-                try {
-                    valA = a.opt("skill_rating");
-                    valB = b.opt("skill_rating");
-                    if (valA == null || !((valA instanceof JSONObject))) valA = new JSONObject().put("stars", new JSONObject().put("avg_star", 0.0f));
-                    if (valB == null || !((valB instanceof JSONObject))) valB = new JSONObject().put("stars", new JSONObject().put("avg_star", 0.0f));
-
-                    JSONObject starsAObject = ((JSONObject) valA).getJSONObject("stars");
-                    JSONObject starsBObject = ((JSONObject) valB).getJSONObject("stars");
-                    int starsA = starsAObject.has("total_star") ? starsAObject.getInt("total_star") : 0;
-                    int starsB = starsBObject.has("total_star") ? starsBObject.getInt("total_star") : 0;
-
-                    float avgA = starsAObject.getFloat("avg_star");
-                    float avgB = starsBObject.getFloat("avg_star");
-
-                    if ((starsA < 10 && starsB < 10) || (starsA >= 10 && starsB >= 10)) {
-                        result = ascending ? Float.compare(avgA, avgB) : Float.compare(avgB, avgA);
-                    } else if (starsA < 10) {
-                        return ascending ? -1 : 1;
-                    } else {
-                        return ascending ? 1 : -1;
-                    }
-                } catch (Exception e) {
-                    e.printStackTrace();
-                }
-                return result;
-            }
-        });
-    }
-
-    public static void sortByCreationTime(List<JSONObject> jsonValues, boolean ascending) {
-        Collections.sort(jsonValues, new Comparator<JSONObject>() {
-            private static final String KEY_NAME = "creationTime";
-            @Override
-            public int compare(JSONObject a, JSONObject b) {
-                String valA = new String();
-                String valB = new String();
-                int result = 0;
-
-                try {
-                    valA = a.get(KEY_NAME).toString();
-                    valB = b.get(KEY_NAME).toString();
-                    result = ascending ? valA.compareToIgnoreCase(valB) : valB.compareToIgnoreCase(valA);
-                } catch (JSONException e) {
-                    e.printStackTrace();
-                }
-                return result;
-            }
-        });
-    }
-
-    public static void sortByModifiedTime(List<JSONObject> jsonValues, boolean ascending) {
-        Collections.sort(jsonValues, new Comparator<JSONObject>() {
-            private static final String KEY_NAME = "lastModifiedTime";
-            @Override
-            public int compare(JSONObject a, JSONObject b) {
-                String valA = new String();
-                String valB = new String();
-                int result = 0;
-
-                try {
-                    valA = a.get(KEY_NAME).toString();
-                    valB = b.get(KEY_NAME).toString();
-                    result = ascending ? valA.compareToIgnoreCase(valB) : valB.compareToIgnoreCase(valA);
-                } catch (JSONException e) {
-                    e.printStackTrace();
-                }
-                return result;
-            }
-        });
-    }
-
-    public static void sortBySkillName(List<JSONObject> jsonValues, boolean ascending) {
-        Collections.sort(jsonValues, new Comparator<JSONObject>() {
-            private static final String KEY_NAME = "skill_name";
-            @Override
-            public int compare(JSONObject a, JSONObject b) {
-                String valA = new String();
-                String valB = new String();
-
-                try {
-                    valA = a.get(KEY_NAME).toString();
-                    valB = b.get(KEY_NAME).toString();
-                } catch (JSONException e) {
-                    //do nothing
-                }
-                return ascending ? valA.compareToIgnoreCase(valB) : valB.compareToIgnoreCase(valA);
-            }
-        });
-    }
-
-    public static void sortByUsageCount(List<JSONObject> jsonValues, boolean ascending) {
-        Collections.sort(jsonValues, new Comparator<JSONObject>() {
-            @Override
-            public int compare(JSONObject a, JSONObject b) {
-                int valA;
-                int valB;
-                int result=0;
-
-                try {
-                    valA = a.getInt("usage_count");
-                    valB = b.getInt("usage_count");
-                    result = ascending ? Integer.compare(valA, valB) : Integer.compare(valB, valA);
-                } catch (JSONException e) {
-                    e.printStackTrace();
-                }
-                return result;
-            }
-        });
-    }
-
-    public static void sortByFeedbackCount(List<JSONObject> jsonValues, boolean ascending) {
-        Collections.sort(jsonValues, new Comparator<JSONObject>() {
-            @Override
-            public int compare(JSONObject a, JSONObject b) {
-                Object valA, valB;
-                int result=0;
-
-                try {
-                    valA = a.opt("skill_rating");
-                    valB = b.opt("skill_rating");
-                    if (valA == null || !(valA instanceof JSONObject) || ((JSONObject) valA).opt("feedback_count") == null) valA = new JSONObject().put("feedback_count", 0);
-                    if (valB == null || !(valB instanceof JSONObject) || ((JSONObject) valB).opt("feedback_count") == null) valB = new JSONObject().put("feedback_count", 0);
-
-                    result = ascending ?
-                            Integer.compare(
-                                    ((JSONObject) valA).getInt("feedback_count"),
-                                    ((JSONObject) valB).getInt("feedback_count")) :
-                            Integer.compare(
-                                    ((JSONObject) valB).getInt("feedback_count"),
-                                    ((JSONObject) valA).getInt("feedback_count")
-                    );
-                } catch (JSONException e) {
-                    e.printStackTrace();
-                }
-                return result;
-            }
-        });
-    }
-
-    public static boolean getSkillReviewStatus(String model, String group, String language, String skillname) {
-        // skill status
-        JsonTray skillStatus = DAO.skillStatus;
-        if (skillStatus.has(model)) {
-            JSONObject modelName = skillStatus.getJSONObject(model);
-            if (modelName.has(group)) {
-                JSONObject groupName = modelName.getJSONObject(group);
-                if (groupName.has(language)) {
-                    JSONObject languageName = groupName.getJSONObject(language);
-                    if (languageName.has(skillname)) {
-                        JSONObject skillName = languageName.getJSONObject(skillname);
-
-                        if (skillName.has("reviewed")) {
-                            return true;
-                        }
-                    }
-                }
-            }
-        }
-        return false;
-    }
-
-    public static boolean getSkillEditStatus(String model, String group, String language, String skillname) {
-        // skill status
-        JsonTray skillStatus = DAO.skillStatus;
-        if (skillStatus.has(model)) {
-            JSONObject modelName = skillStatus.getJSONObject(model);
-            if (modelName.has(group)) {
-                JSONObject groupName = modelName.getJSONObject(group);
-                if (groupName.has(language)) {
-                    JSONObject languageName = groupName.getJSONObject(language);
-                    if (languageName.has(skillname)) {
-                        JSONObject skillName = languageName.getJSONObject(skillname);
-
-                        if (skillName.has("editable")) {
-                            return false;
-                        }
-                    }
-                }
-            }
-        }
-        return true;
-    }
-
-    public static boolean isStaffPick(String model, String group, String language, String skillname) {
-        // return true if skill is a staff pick
-        JsonTray skillStatus = DAO.skillStatus;
-        if (skillStatus.has(model)) {
-            JSONObject modelName = skillStatus.getJSONObject(model);
-            if (modelName.has(group)) {
-                JSONObject groupName = modelName.getJSONObject(group);
-                if (groupName.has(language)) {
-                    JSONObject languageName = groupName.getJSONObject(language);
-                    if (languageName.has(skillname)) {
-                        JSONObject skillName = languageName.getJSONObject(skillname);
-
-                        if (skillName.has("staffPick")) {
-                            return true;
-                        }
-                    }
-                }
-            }
-        }
-        return false;
-    }
-
-    public static boolean isSystemSkill(String model, String group, String language, String skillname) {
-        // return true if skill is a system skill
-        JsonTray skillStatus = DAO.skillStatus;
-        if (skillStatus.has(model)) {
-            JSONObject modelName = skillStatus.getJSONObject(model);
-            if (modelName.has(group)) {
-                JSONObject groupName = modelName.getJSONObject(group);
-                if (groupName.has(language)) {
-                    JSONObject languageName = groupName.getJSONObject(language);
-                    if (languageName.has(skillname)) {
-                        JSONObject skillName = languageName.getJSONObject(skillname);
-
-                        if (skillName.has("systemSkill")) {
-                            return true;
-                        }
-                    }
-                }
-            }
-        }
-        return false;
-    }
-
-    public static int getSkillUsage(String model, String group, String language, String skillname, int duration) {
-        // usage
-        int totalUsage = 0;
-        JsonTray skillUsage = DAO.skillUsage;
-        if (skillUsage.has(model)) {
-            JSONObject modelName = skillUsage.getJSONObject(model);
-            if (modelName.has(group)) {
-                JSONObject groupName = modelName.getJSONObject(group);
-                if (groupName.has(language)) {
-                    JSONObject languageName = groupName.getJSONObject(language);
-                    if (languageName.has(skillname)) {
-                        JSONArray skillName = languageName.getJSONArray(skillname);
-
-                        // Fetch skill usage data for sorting purpose.
-                        int startIndex = 0;
-                        if(duration >= 0) {
-                            startIndex = skillName.length() >= duration ? skillName.length()-duration : 0;
-                        }
-                        for (int i = startIndex; i<skillName.length(); i++)
-                        {
-                            JSONObject dayUsage = skillName.getJSONObject(i);
-                            totalUsage += dayUsage.getInt("count");
-                        }
-                        return totalUsage;
-                    }
-                    else {
-                        return 0;
-                    }
-                }
-            }
-        }
-        return 0;
-    }
-
-    public static JSONObject readJsonSkill(File file) throws JSONException, FileNotFoundException {
-        JSONObject json = new JSONObject(new JSONTokener(new FileReader(file)));
-        //System.out.println(json.toString(2)); // debug
-        return json;
     }
 
     public void setAuthor(String author) {
