+++ src/ai/susi/server/api/cms/ExampleSkillService.java
@@ -26,7 +26,6 @@ import ai.susi.server.*;
 import org.json.JSONObject;
 
 import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
 import java.util.Map;
 
 /**
