+++ src/ai/susi/server/api/cms/FeedbackLogService.java
@@ -22,14 +22,12 @@ package ai.susi.server.api.cms;
 import ai.susi.DAO;
 import ai.susi.json.JsonObjectWithDefault;
 import ai.susi.json.JsonTray;
-import ai.susi.mind.SusiSkill;
 import ai.susi.server.*;
 import org.json.JSONArray;
 import org.json.JSONObject;
 
 import javax.servlet.http.HttpServletResponse;
 import java.io.File;
-import java.io.IOException;
 import java.sql.Timestamp;
 
 
@@ -69,7 +67,7 @@ public class FeedbackLogService extends AbstractAPIHandler implements APIHandler
         String language_name = call.get("language", "en");
         File language = new File(group, language_name);
         String skill_name = call.get("skill", null);
-        File skill = SusiSkill.getSkillFileInLanguage(language, skill_name, false);
+        File skill = DAO.getSkillFileInLanguage(language, skill_name, false);
         String skill_rate = call.get("rating", null);
         String user_query = call.get("user_query", null);
         String susi_reply = call.get("susi_reply", null);
